feat: Add new house scene with textures and collision shapes
This commit introduces a new 'House' scene in the 'scenes/houses/' directory, using the 'house.tscn' file. The scene comes with multiple texture and collision shape definitions, modeled as a 2D area in the game environment. This includes settings for physics layers and angular velocity for objects within the 'House' scene. A  TileMap node for the interior and exterior walls of the house is also added.